Examples of the behavior of the "include" keyword in ruby
This is a simple spec test that demonstrates some things about how "include" works in ruby.  Specifically, it's intended to demonstrate the case where "include" is being used to basically import some static methods / functions from another namespace.  In this case, the methods in question are stateless and don't really relate to the object heirarchy at all (they might be called "static methods" in java, or "pure functions" in other contexts), yet we are still creating an is-a relationship to the containing namespace in order to provide access to them.  This can make it tricky to maintain code, as it results in superfluous and bloated object inheritance heirarchies, and can mask the details about where a method is defined.